refactor(cache): split cache.go into focused domain files
Move key computation functions (EngineKey, GameServerKey, GameClientKey,
ContainerKey, hash) to cache_keys.go and filesystem/git helpers (gitHEAD,
fileKey, dirManifest) to cache_helpers.go. cache.go retains only the core
types (StageKey, Entry, Cache), persistence (Load, Save), and high-level
pipeline functions (IsHit, MissReason, Set, CheckSkip, RecordBuild).
No public API or behavior changes.
